Abstract Out MultiCheckTable Implementation
